move: fitness data model as widgets
