+++ apps/web/src/app/App.tsx
@@ -1,8 +1,8 @@
 import { HeroUIProvider, Spinner } from '@heroui/react'
 import { Suspense } from 'react'
 import { useNavigate } from 'react-router-dom'
-import FitnessDataModal from '@globalUi/FitnessDataModal'
 import useTheme from '@shared/hooks/useTheme'
+import FitnessDataModal from '@widgets/FitnessDataModal'
 import AlertModal from '@widgets/alert'
 import BottomNavi from '@widgets/bottomNavi'
 import DefaultLayout from './layout/DefaultLayout'
