+++ apps/web/src/components/ui/Fitness/FitnessList.tsx
@@ -1,6 +1,6 @@
 import { Suspense } from 'react';
-import useFitnessDataModal from '@globalUi/FitnessDataModal/hooks/useFitnessDataModal';
 import useSpinner from '@shared/hooks/useSpinner';
+import useFitnessDataModal from '@widgets/FitnessDataModal/hooks/useFitnessDataModal';
 import FitnessItem from './FitnessItem';
 
 export type FitnessListSelectedProps = { selected: boolean, idx: number }
