+++ apps/web/src/widgets/FitnessDataModal/components/DisplayExerciseFinishHistory.tsx
