+++ apps/web/src/widgets/FitnessDataModal/components/FitnessDataModalContent.tsx
