+++ apps/web/src/widgets/FitnessDataModal/components/FitnessPreviewVideo.ts
